requirements.txt
Django==1.7.7
Markdown==2.6.2
Pillow==2.8.1
South==1.0.2
Unidecode==0.04.17
YURL==0.13
aldryn-boilerplates==0.6
argparse==1.2.1
beautifulsoup4==4.3.2
bleach==1.4.1
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-appdata==0.1.4
django-blog-zinnia==0.15.1
django-bootstrap-form==3.2
django-ckeditor==5.1.1
django-classy-tags==0.6.1
django-cms==3.0.12
django-contrib-comments==1.6.0
django-filer==0.9.9
django-geoposition==0.3.0
django-haystack==2.3.1
django-hvad==1.2.0
django-imagekit==3.3
django-mptt==0.6.1
django-polymorphic==0.6.1
django-sekizai==0.8.1
django-session-security==2.2.3
django-sortedm2m==0.9.4
django-storages==1.1.8
django-suit==0.2.21
django-tagging==0.3.4
django-taggit==0.11.2
django-xmlrpc==0.1.5
djangocms-admin-style==0.2.5
easy-thumbnails==2.2
filebrowser-safe==0.3.8
future==0.9.0
grappelli-safe==0.3.13
gunicorn==19.3.0
html5lib==0.999
oauthlib==0.7.2
pilkit==1.1.13
psycopg2==2.6
pyparsing==2.0.3
pytz==2015.2
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
static3==0.5.1
textile==2.2.2
tzlocal==1.0
wsgiref==0.1.2